Refactored framework code
