+++ HW8/app/db.json
@@ -1,0 +1,12 @@
+{
+    "toys": [
+        {
+            "name": "fasdf",
+            "price": 100.0,
+            "age_range": [
+                "10",
+                "90"
+            ]
+        }
+    ]
+}
