+++ HW8/app/templates.py
@@ -1,9 +1,5 @@
-import os
+from app.settings import TEMPLATE_DIR
 from jinja2 import Environment, FileSystemLoader, Template
-
-BASE_DIR = os.path.abspath(os.path.dirname(__file__))
-TEMPLATE_DIR = "templates/"
-TEMPLATES_DIR = os.path.join(BASE_DIR, TEMPLATE_DIR)
 
 jinja_env = Environment(loader = FileSystemLoader(TEMPLATE_DIR))
